Adding development profile bean
